# Conflicts:
#	lib/main.dart
#	lib/models/task.dart
#	lib/utils/scheduler.dart
#	lib/utils/task_urgency_calculator.dart

+++ lib/models/scheduled_task.dart
@@ -18,7 +18,7 @@ class ScheduledTask extends HiveObject {
   DateTime endTime;
 
   @HiveField(3)
-  int urgency;
+  double urgency;
 
   @HiveField(4)
   ScheduledTaskType type;
